+++ VoiceControl/classes/Tools/sqlite3/SQLiteWrapper.h
@@ -85,7 +85,7 @@ public:
 	
 	bool Open(std::string const& db_file);
 	
-    int lastInsertRowid();
+    long long lastInsertRowid();
     
 	class ResultRecord {
 	public:
